✨ Created context for pokemons
