+++ src/hooks/usePokemons.tsx
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { PokemonContext } from '../context/pokemonContext';
+
+export const usePokemon = () => useContext(PokemonContext);
