+++ src/pages/_app.tsx
@@ -2,6 +2,7 @@ import { AppProps } from 'next/app';
 import { ThemeProvider } from 'styled-components';
 
 import FavoritesProvider from '../context/favoriteContext';
+import PokemonProvider from '../context/pokemonContext';
 
 import GlobalStyles from '../styles/global';
 import theme from '../styles/theme';
@@ -11,9 +12,11 @@ function MyApp({ Component, pageProps }: AppProps) {
     <ThemeProvider theme={theme}>
       <GlobalStyles />
 
-      <FavoritesProvider>
-        <Component {...pageProps} />
-      </FavoritesProvider>
+      <PokemonProvider>
+        <FavoritesProvider>
+          <Component {...pageProps} />
+        </FavoritesProvider>
+      </PokemonProvider>
     </ThemeProvider>
   );
 }
