+++ src/pages/dashboard/favorites.tsx
